Move action to a separate controller
